MobileUI Picking — split the whole-HU over-pick RED into its own spec file
The whole-HU scenario needs a BOM/manufacturing masterdata builder unrelated to the
LU/TU and LU/CU builders already in overPickingPrompt.spec.js. e2e/CLAUDE.md requires a
spec file to read as "one setup, N tests exercising it", so the fixture and its scenario
move to overPickingPrompt_wholeHU.spec.js. overPickingPrompt.spec.js returns to its
pre-RED content.
